+++ src/RTUutils.h
@@ -59,11 +59,11 @@ protected:
   static int UARTinit(HardwareSerial& serial, int thresholdBytes = 1);
 
 // receive: get a Modbus message from serial, maintaining timeouts etc.
-  static ModbusMessage receive(HardwareSerial& serial, uint32_t timeout, uint32_t& lastMicros, uint32_t interval);
+  static ModbusMessage receive(HardwareSerial& serial, uint32_t timeout, unsigned long& lastMicros, uint32_t interval);
 
 // send: send a Modbus message in either format (ModbusMessage or data/len)
-  static void send(HardwareSerial& serial, uint32_t& lastMicros, uint32_t interval, RTScallback r, const uint8_t *data, uint16_t len);
-  static void send(HardwareSerial& serial, uint32_t& lastMicros, uint32_t interval, RTScallback r, ModbusMessage raw);
+  static void send(HardwareSerial& serial, unsigned long& lastMicros, uint32_t interval, RTScallback r, const uint8_t *data, uint16_t len);
+  static void send(HardwareSerial& serial, unsigned long& lastMicros, uint32_t interval, RTScallback r, ModbusMessage raw);
 };
 
 #endif
